同上

+++ install.sh
@@ -12,28 +12,43 @@ elif [ $sysOS == "Linux" ];then
 
     # 安装cmake
     #wget https://github.com/Kitware/CMake/releases/download/v3.15.0-rc3/cmake-3.15.0-rc3-Linux-x86_64.tar.gz
-    sudo apt autoremove cmake -y
-    cd pkg
-    sudo tar -zxvf cmake-3.15.0-rc3-Linux-x86_64.tar.gz -C /
-    sudo rm -rf /cmake
-    sudo mv /cmake-3.15.0-rc3-Linux-x86_64 /cmake
-    sudo ln -sf /cmake/bin/* /usr/bin/
-    cd $runPath
+    cmakePath="/cmake_3.15.0"
+    if [ ! -d $cmakePath ]; then
+        sudo apt autoremove cmake -y
+        cd pkg
+        sudo tar -zxvf cmake-3.15.0-rc3-Linux-x86_64.tar.gz -C /
+        sudo rm -rf $cmakePath
+        sudo mv /cmake-3.15.0-rc3-Linux-x86_64 $cmakePath
+        sudo ln -sf $cmakePath/bin/* /usr/bin/
+        cd $runPath
+    fi
 
 
 fi
 
 # 安装flatbuffers
 #git clone https://github.com/google/flatbuffers.git
-cd pkg/flatbuffers
-mkdir release && cd release
-cmake ..
-make && sudo make install
-cd .. && rm -rf release
-cd $runPath
+fbsVersion="null"
+if type flatc >/dev/null 2>&1;then 
+    fbsVersion=$(flatc --version)
+fi
+echo $fbsVersion
+if [ "$fbsVersion" != "flatc version 1.11.0" ];then
+    cd pkg/flatbuffers
+    rm -rf release
+    mkdir release && cd release
+    cmake ..
+    make && sudo make install
+    cd .. && rm -rf release
+    cd $runPath
+fi
 
 # 安装程序
-sudo rm -rf build && mkdir build && cd build
+cd $runPath
+if [ ! -d "build" ]; then
+    mkdir build
+fi
+cd build
 cmake ..
 make
 sudo make install
